Moved WindowEvents to API package.
Created AbstractWindow.  VanillaRenderMaterials is also now in the API package as it's a static class and seems to be universally necessary.

Updated Docs for InventoryCanSetevent.  removed PlayerSlotChangeEvent, it was unnecessary to begin with.
