+++ src/main/java/org/spout/vanilla/api/data/VanillaRenderMaterials.java
@@ -24,7 +24,7 @@
  * License and see <http://spout.in/licensev1> for the full license, including
  * the MIT license.
  */
-package org.spout.vanilla.plugin.data;
+package org.spout.vanilla.api.data;
 
 import org.spout.api.FileSystem;
 import org.spout.api.Spout;
