+++ src/main/java/org/spout/vanilla/api/event/window/WindowCloseEvent.java
@@ -24,16 +24,16 @@
  * License and see <http://spout.in/licensev1> for the full license, including
  * the MIT license.
  */
-package org.spout.vanilla.plugin.event.window;
+package org.spout.vanilla.api.event.window;
 
 import org.spout.api.event.HandlerList;
 
-import org.spout.vanilla.plugin.inventory.window.Window;
+import org.spout.vanilla.api.inventory.window.AbstractWindow;
 
 public class WindowCloseEvent extends WindowEvent {
 	private static HandlerList handlers = new HandlerList();
 
-	public WindowCloseEvent(Window window) {
+	public WindowCloseEvent(AbstractWindow window) {
 		super(window);
 	}
