+++ src/main/java/org/spout/vanilla/api/event/window/WindowEvent.java
@@ -24,25 +24,25 @@
  * License and see <http://spout.in/licensev1> for the full license, including
  * the MIT license.
  */
-package org.spout.vanilla.plugin.event.window;
+package org.spout.vanilla.api.event.window;
 
-import org.spout.api.event.HandlerList;
+import org.spout.api.event.Event;
+import org.spout.api.protocol.event.ProtocolEvent;
 
-import org.spout.vanilla.plugin.inventory.window.Window;
+import org.spout.vanilla.api.inventory.window.AbstractWindow;
 
-public class WindowOpenEvent extends WindowEvent {
-	private static HandlerList handlers = new HandlerList();
+public abstract class WindowEvent extends Event implements ProtocolEvent {
+	private AbstractWindow window;
 
-	public WindowOpenEvent(Window window) {
-		super(window);
+	public WindowEvent(AbstractWindow window) {
+		this.window = window;
 	}
 
-	@Override
-	public HandlerList getHandlers() {
-		return handlers;
-	}
-
-	public static HandlerList getHandlerList() {
-		return handlers;
+	/**
+	 * Gets the Window of this Event
+	 * @return the Window
+	 */
+	public AbstractWindow getWindow() {
+		return this.window;
 	}
 }
