+++ src/main/java/org/spout/vanilla/plugin/component/misc/HealthComponent.java
@@ -75,7 +75,7 @@ import org.spout.vanilla.plugin.component.substance.XPOrb;
 import org.spout.vanilla.plugin.component.substance.object.Item;
 import org.spout.vanilla.plugin.configuration.VanillaConfiguration;
 import org.spout.vanilla.plugin.data.VanillaData;
-import org.spout.vanilla.plugin.data.VanillaRenderMaterials;
+import org.spout.vanilla.api.data.VanillaRenderMaterials;
 import org.spout.vanilla.plugin.protocol.msg.entity.EntityStatusMessage;
 
 /**
