+++ src/main/java/org/spout/vanilla/plugin/component/player/hud/VanillaArmorWidget.java
@@ -33,7 +33,7 @@ import org.spout.api.gui.component.RenderPartsHolderComponent;
 import org.spout.api.gui.render.RenderPart;
 import org.spout.api.math.Rectangle;
 import org.spout.vanilla.plugin.component.player.HUDComponent;
-import org.spout.vanilla.plugin.data.VanillaRenderMaterials;
+import org.spout.vanilla.api.data.VanillaRenderMaterials;
 
 public class VanillaArmorWidget extends ArmorWidget {
 	@Override
