+++ src/main/java/org/spout/vanilla/plugin/component/player/hud/VanillaExpBar.java
@@ -37,7 +37,7 @@ import org.spout.api.gui.render.RenderPart;
 import org.spout.api.math.Rectangle;
 
 import org.spout.vanilla.plugin.component.player.HUDComponent;
-import org.spout.vanilla.plugin.data.VanillaRenderMaterials;
+import org.spout.vanilla.api.data.VanillaRenderMaterials;
 
 public class VanillaExpBar extends ExpBarWidget {
 	@Override
