+++ src/main/java/org/spout/vanilla/plugin/component/player/hud/VanillaHotBar.java
@@ -36,7 +36,7 @@ import org.spout.api.math.Rectangle;
 import org.spout.vanilla.api.inventory.entity.QuickbarInventory;
 import org.spout.vanilla.plugin.component.inventory.PlayerInventory;
 import org.spout.vanilla.plugin.component.player.HUDComponent;
-import org.spout.vanilla.plugin.data.VanillaRenderMaterials;
+import org.spout.vanilla.api.data.VanillaRenderMaterials;
 
 public class VanillaHotBar extends HotBarWidget {
 	@Override
