+++ src/main/java/org/spout/vanilla/plugin/inventory/window/DefaultWindow.java
@@ -37,11 +37,13 @@ import org.spout.api.inventory.util.GridIterator;
 import org.spout.api.math.Vector2;
 import org.spout.api.plugin.Platform;
 
+import org.spout.vanilla.api.component.inventory.PlayerInventoryComponent;
 import org.spout.vanilla.api.inventory.entity.ArmorInventory;
-import org.spout.vanilla.plugin.inventory.player.PlayerCraftingInventory;
+import org.spout.vanilla.api.inventory.window.ClickArguments;
+import org.spout.vanilla.api.inventory.window.WindowType;
 
-import org.spout.vanilla.plugin.component.inventory.PlayerInventory;
 import org.spout.vanilla.plugin.inventory.entity.EntityArmorInventory;
+import org.spout.vanilla.plugin.inventory.player.PlayerCraftingInventory;
 import org.spout.vanilla.plugin.inventory.util.InventoryConverter;
 
 public class DefaultWindow extends Window {
@@ -49,7 +51,7 @@ public class DefaultWindow extends Window {
 
 	public DefaultWindow(Player owner) {
 		super(owner, WindowType.DEFAULT, "Inventory", 9);
-		PlayerInventory inventory = getPlayerInventory();
+		PlayerInventoryComponent inventory = getPlayerInventory();
 
 		addInventoryConverter(new InventoryConverter(inventory.getArmor(), "8, 7, 6, 5", new Vector2[]{
 				Vector2.ZERO, Vector2.ZERO, Vector2.ZERO, Vector2.ZERO
@@ -62,7 +64,7 @@ public class DefaultWindow extends Window {
 
 	@Override
 	public void close() {
-		PlayerInventory pInv = getPlayerInventory();
+		PlayerInventoryComponent pInv = getPlayerInventory();
 		//Disconnecting
 		if (pInv == null) {
 			return;
@@ -97,7 +99,7 @@ public class DefaultWindow extends Window {
 		if (Spout.getPlatform() == Platform.CLIENT) {
 			throw new IllegalStateException("Shift click handling is handled server side.");
 		}
-		final PlayerInventory inventory = getPlayerInventory();
+		final PlayerInventoryComponent inventory = getPlayerInventory();
 
 		// Transferring to the armor slots
 		if (!(from instanceof EntityArmorInventory)) {
