+++ src/main/java/org/spout/vanilla/plugin/inventory/window/block/CraftingTableWindow.java
@@ -34,7 +34,7 @@ import org.spout.api.math.Vector2;
 import org.spout.vanilla.plugin.inventory.block.CraftingTableInventory;
 import org.spout.vanilla.plugin.inventory.util.InventoryConverter;
 import org.spout.vanilla.plugin.inventory.window.Window;
-import org.spout.vanilla.plugin.inventory.window.WindowType;
+import org.spout.vanilla.api.inventory.window.WindowType;
 
 public class CraftingTableWindow extends Window {
 	private final CraftingTableInventory inventory = new CraftingTableInventory();
