+++ src/main/java/org/spout/vanilla/plugin/inventory/window/block/chest/ChestWindow.java
@@ -32,7 +32,7 @@ import org.spout.api.math.Vector2;
 import org.spout.vanilla.plugin.component.substance.material.chest.Chest;
 import org.spout.vanilla.plugin.inventory.block.ChestInventory;
 import org.spout.vanilla.plugin.inventory.util.GridInventoryConverter;
-import org.spout.vanilla.plugin.inventory.window.WindowType;
+import org.spout.vanilla.api.inventory.window.WindowType;
 
 public class ChestWindow extends AbstractChestWindow {
 	public ChestWindow(Player owner, Chest chest) {
