+++ src/main/java/org/spout/vanilla/plugin/protocol/handler/window/WindowClickHandler.java
@@ -31,7 +31,7 @@ import org.spout.api.protocol.MessageHandler;
 import org.spout.api.protocol.Session;
 
 import org.spout.vanilla.plugin.component.inventory.WindowHolder;
-import org.spout.vanilla.plugin.inventory.window.ClickArguments;
+import org.spout.vanilla.api.inventory.window.ClickArguments;
 import org.spout.vanilla.plugin.inventory.window.Window;
 import org.spout.vanilla.plugin.protocol.msg.window.WindowClickMessage;
 import org.spout.vanilla.plugin.protocol.msg.window.WindowTransactionMessage;
